Implement InventorySnapshots to manage inventory and manually setting amount
